// lib/prisma/enums.ts
export const ROLE = {
  SYSTEM_OWNER: "SYSTEM_OWNER",
  SYSTEM_ADMIN: "SYSTEM_ADMIN",
  SYSTEM_USER: "SYSTEM_USER",
  USER: "USER",
} as const;

export type ROLE = (typeof ROLE)[keyof typeof ROLE];
// ...
export const AUDIT_LOG_ACTION = {
  USER_REGISTERED: "USER_REGISTERED",
  USER_LOGGED_IN: "USER_LOGGED_IN",
  USER_LOGGED_OUT: "USER_LOGGED_OUT",
  USER_RESEND_VERIFICATION: "USER_RESEND_VERIFICATION",
  USER_FORGOT_PASSWORD: "USER_FORGOT_PASSWORD",
  USER_RESET_PASSWORD: "USER_RESET_PASSWORD",
  USER_VERIFIED: "USER_VERIFIED",
  USER_SUSPENDED: "USER_SUSPENDED",
  USER_UNSUSPENDED: "USER_UNSUSPENDED",
  USER_APPROVED: "USER_APPROVED",
  USER_PASSWORD_CHANGED: "USER_PASSWORD_CHANGED",
  USER_CREATED: "USER_CREATED",
  USER_UPDATED: "USER_UPDATED",
  USER_DELETED: "USER_DELETED",
  SESSION_CREATED: "SESSION_CREATED",
  SESSION_REVOKED: "SESSION_REVOKED",
  SYSTEM_ACCESS_CREATED: "SYSTEM_ACCESS_CREATED",
  SYSTEM_ACCESS_UPDATED: "SYSTEM_ACCESS_UPDATED",
  SYSTEM_ACCESS_SUSPENDED: "SYSTEM_ACCESS_SUSPENDED",
  SYSTEM_ACCESS_UNSUSPENDED: "SYSTEM_ACCESS_UNSUSPENDED",
  SYSTEM_ACCESS_DELETED: "SYSTEM_ACCESS_DELETED",
} as const;

export type AUDIT_LOG_ACTION =
  (typeof AUDIT_LOG_ACTION)[keyof typeof AUDIT_LOG_ACTION];

export const AUDIT_LOG_ENTITY = {
  USER: "USER",
  SESSION: "SESSION",
  SYSTEM_ACCESS: "SYSTEM_ACCESS",
  WORKSPACE_ACCOUNT: "WORKSPACE_ACCOUNT",
  RELEASE: "RELEASE",
  TRACK: "TRACK",
  VIDEO: "VIDEO",
  RINGTONE: "RINGTONE",
  ARTIST: "ARTIST",
  PERFORMER: "PERFORMER",
  PRODUCER_AND_ENGINEER: "PRODUCER_AND_ENGINEER",
  ARTIST_PERFORMER_PRODUCER_AND_ENGINEER:
    "ARTIST_PERFORMER_PRODUCER_AND_ENGINEER",
  WRITER: "WRITER",
  PUBLISHER: "PUBLISHER",
  LABEL: "LABEL",
  TRANSACTION: "TRANSACTION",
  WITHDRAWAL: "WITHDRAWAL",
  SHARED_WORKSPACE_ACCOUNT_ACCESS: "SHARED_WORKSPACE_ACCOUNT_ACCESS",
  RIGHTS_MANAGEMENT: "RIGHTS_MANAGEMENT",
} as const;

export type AUDIT_LOG_ENTITY =
  (typeof AUDIT_LOG_ENTITY)[keyof typeof AUDIT_LOG_ENTITY];